Header refactoring: unify and de-catchall rtree module.
