+++ test/unit/spin.c
@@ -1,5 +1,7 @@
 #include "test/jemalloc_test.h"
 
+#include "jemalloc/internal/spin.h"
+
 TEST_BEGIN(test_spin) {
 	spin_t spinner = SPIN_INITIALIZER;
